# Conflicts:
#	client/pom.xml
#	client/src/test/java/pt/ulisboa/tecnico/sec/g19/hdscoin/client/RegisterTest.java
#	pom.xml

+++ common/src/main/java/pt/ulisboa/tecnico/sec/g19/hdscoin/common/execeptions/InvalidAmountException.java
@@ -2,7 +2,7 @@ package pt.ulisboa.tecnico.sec.g19.hdscoin.common.execeptions;
 
 public class InvalidAmountException extends Exception {
 
-    public InvalidAmountException(String message, double amount) {
+    public InvalidAmountException(String message, int amount) {
         super(message + " Amount: " + amount);
     }
 }
